python/sentinel-connectors/pyproject.toml
[project]
name = "sentinel-connectors"
version = "0.1.0"
description = "Sentinel Connectors — cloud and identity integration framework"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "sentinel-api",
]

[project.optional-dependencies]
aws = ["boto3>=1.35.0"]
azure = ["azure-identity>=1.19.0", "azure-mgmt-compute>=33.0.0", "azure-mgmt-network>=28.0.0", "msgraph-sdk>=1.12.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "moto[ec2,iam,s3]>=5.0",
    "ruff>=0.9",
    "boto3>=1.35.0",
]

[tool.uv.sources]
sentinel-api = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentinel_connectors"]
